OmDet/pyproject.toml
[project]
name = "omdet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "modelscope>=1.36.3",
    "pillow>=12.2.0",
    "transformers>=5.7.0",
]
